requirements.txt
feedparser>=6.0.0
requests>=2.25.0
beautifulsoup4>=4.9.3
python-dateutil>=2.8.1
youtube-transcript-api>=0.5.0 
demjson3>=3.0.0